requirements.txt
streamlit==1.26.0
sounddevice==0.4.6
numpy==1.26.0
scipy==1.11.3
python-docx==0.8.11
openai==0.30.0
faiss-cpu==1.7.4
python-dotenv==1.0.0
langchain==0.0.330
langchain-community==0.0.14
